// src/app/models/enums/bookActions.ts
export enum BookActionTypes {
    GetBooks = '[Book] Get Books',
    GetBooksSuccess = '[Book] Get Books Success',
    GetBooksError = '[Book] Get Books Error',

    AddBook = '[Book] Add Book',
    AddBookSuccess = '[Book] Add Book Success',
    AddBookError = '[Book] Add Book Error',

    UpdateBook = '[Book] Update Book',
    UpdateBookSuccess = '[Book] Update Book Success',
    UpdateBookError = '[Book] Update Book Error',

    DeleteBook = '[Book] Delete Book',
    DeleteBookSuccess = '[Book] Delete Book Success',
    DeleteBookError = '[Book] Delete Book Error',

    SetSelectedBook = '[Book] Set Selected Book',
    ClearSelectedBook = '[Book] Clear Selected Book'
}